requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
git+https://github.com/openai/CLIP.git
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
exceptiongroup==1.2.0
fastapi==0.109.2
ffmpy==0.3.1
filelock==3.13.1
fonttools==4.48.1
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.1.3
gradio==3.47.1
gradio-client==0.6.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.3
idna==3.6
importlib-resources==6.1.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kornia==0.7.1
linkify-it-py==2.0.3
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.8.2
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
orjson==3.9.13
packaging==23.2
pandas==2.2.0
pillow==10.2.0
pydantic==2.6.1
pydantic-core==2.16.2
pydub==0.25.1
pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-multipart==0.0.7
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
ruff==0.2.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
sympy==1.12
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.1
triton==2.2.0
typer==0.9.0
typing-extensions==4.9.0
tzdata==2023.4
uc-micro-py==1.0.3
urllib3==2.2.0
uvicorn==0.27.0.post1
wcwidth==0.2.13
websockets==11.0.3
yarl==1.9.4
zipp==3.17.0
